Cleanup

+++ src/sys/compat/ndis/kern_windrv.c
@@ -301,7 +301,7 @@ windrv_unload(module_t mod, vm_offset_t img, int len)
  * calls the driver's DriverEntry() routine.
  */
 int
-windrv_load(module_t mod, vm_offset_t img, int len, interface_type bustype,
+windrv_load(module_t mod, vm_offset_t img, size_t len, interface_type bustype,
     void *devlist, ndis_cfg *regvals)
 {
 	image_import_descriptor imp_desc;
